requirements.txt
django
PyMySQL
mysqlclient
djangorestframework
django-cors-headers